+++ src/Apis/WeddingInvitationApi.js
@@ -1,10 +1,13 @@
 import instance from "./instance";
 
-class WeddingInvitationApi  {
-    static update(data) {
-        return instance.put('/api/v1/login', data)
+class WeddingInvitationApi {
+    static getWeddingInfo() {
+        return instance.get('/api/v1/wedding_invitation');
     }
-   
+
+    static updateWeddingInfo(data) {
+        return instance.put('/api/v1/wedding_invitation', data);
+    }
 }
 
 export default WeddingInvitationApi;
