+++ src/Component/Auth/FormLogin.js
@@ -13,7 +13,7 @@ function FormLogin({ formik }) {
           id="email"
           type="email"
           autoComplete="email"
-          className="mt-1 w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-rose-300"
+          className="mt-1 w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-indigo-500"
           value={formik.values.email}
           onChange={formik.handleChange}
           onBlur={formik.handleBlur}
@@ -34,7 +34,7 @@ function FormLogin({ formik }) {
           id="password"
           type="password"
           autoComplete="password"
-          className="mt-1 w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-rose-300"
+          className="mt-1 w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-indigo-500"
           value={formik.values.password}
           onChange={formik.handleChange}
           onBlur={formik.handleBlur}
