+++ src/Component/sections/wedding-info/InputField.js
@@ -1,21 +1,21 @@
 import React from 'react';
 
-const TextAreaField = ({ label, icon, name, formik, placeholder, rows = 3 }) => {
+const InputField = ({ color, label, icon, name, formik, placeholder, type = 'text' }) => {
     return (
         <div>
             <label className="text-sm font-medium text-gray-700 mb-1 flex items-center gap-1">
                 {icon}
                 {label}
             </label>
-            <textarea
+            <input
+                type={type}
                 name={name}
                 value={formik.values[name]}
                 onChange={formik.handleChange}
                 onBlur={formik.handleBlur}
-                rows={rows}
-                className={`w-full px-3 py-2 border border-gray-300 rounded-md focus:ring-indigo-500 focus:border-indigo-500 ${
-                    formik.touched[name] && formik.errors[name] ? 'border-red-500' : ''
-                }`}
+                className={`w-full px-3 py-2 border border-gray-300 rounded-md outline-none focus:ring-1 
+        ${formik.touched[name] && formik.errors[name] ? 'border-red-500' : ''}
+        ${color === 'indigo' ? 'focus:ring-indigo-500 focus:border-indigo-500' : 'focus:ring-rose-500 focus:border-rose-500'}`}
                 placeholder={placeholder}
             />
             {formik.touched[name] && formik.errors[name] && (
@@ -25,4 +25,4 @@ const TextAreaField = ({ label, icon, name, formik, placeholder, rows = 3 }) => 
     );
 };
 
-export default TextAreaField;
+export default InputField;
