+++ src/Component/sections/wedding-info/TextAreaField.js
@@ -1,21 +1,21 @@
 import React from 'react';
 
-const InputField = ({ label, icon, name, formik, placeholder, type = 'text' }) => {
+const TextAreaField = ({ color, label, icon, name, formik, placeholder, rows = 3 }) => {
     return (
         <div>
             <label className="text-sm font-medium text-gray-700 mb-1 flex items-center gap-1">
                 {icon}
                 {label}
             </label>
-            <input
-                type={type}
+            <textarea
                 name={name}
                 value={formik.values[name]}
                 onChange={formik.handleChange}
                 onBlur={formik.handleBlur}
-                className={`w-full px-3 py-2 border border-gray-300 rounded-md focus:ring-indigo-500 focus:border-indigo-500 ${
-                    formik.touched[name] && formik.errors[name] ? 'border-red-500' : ''
-                }`}
+                rows={rows}
+                className={`w-full px-3 py-2 border outline-none border-gray-300 rounded-md focus:ring-1
+        ${formik.touched[name] && formik.errors[name] ? 'border-red-500' : ''}
+        ${color === 'indigo' ? 'focus:ring-indigo-500 focus:border-indigo-500' : 'focus:ring-rose-500 focus:border-rose-500'}`}
                 placeholder={placeholder}
             />
             {formik.touched[name] && formik.errors[name] && (
@@ -25,4 +25,4 @@ const InputField = ({ label, icon, name, formik, placeholder, type = 'text' }) =
     );
 };
 
-export default InputField;
+export default TextAreaField;
